requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
cassandra-driver
minio
jinja2
aiofiles
python-dotenv
pytest
httpx
